feat: ability generate withSchedule  for boostrap/app.php file
refs: https://app.clickup.com/t/86c4eft6r
