+++ src/Traits/AstValueBuilderTrait.php
@@ -13,47 +13,5 @@ use PhpParser\Node\Scalar\String_;
 
 trait AstValueBuilderTrait
 {
-    protected function makeArg(mixed $value): Arg
-    {
-        list($value) = $this->getPropertyValue($value);
 
-        return new Arg($value);
-    }
-
-    protected function getPropertyValue(mixed $value): array
-    {
-        $type = get_debug_type($value);
-
-        $value = match ($type) {
-            'int' => new Int_($value),
-            'array' => $this->makeArrayValue($value),
-            'string' => new String_($value),
-            'float' => new Float_($value),
-            'bool' => $this->makeBoolValue($value),
-            'null' => new ConstFetch(new Name('null')),
-        };
-
-        return [$value, $type];
-    }
-
-    protected function makeBoolValue(bool $value): ConstFetch
-    {
-        $name = new Name(($value) ? 'true' : 'false');
-
-        return new ConstFetch($name);
-    }
-
-    protected function makeArrayValue(array $values): Array_
-    {
-        $items = [];
-
-        foreach ($values as $key => $val) {
-            list($val) = $this->getPropertyValue($val);
-            list($key) = $this->getPropertyValue($key);
-
-            $items[] = new ArrayItem($val, $key);
-        }
-
-        return new Array_($items);
-    }
 }
