+++ src/Visitors/BaseNodeVisitorAbstract.php
@@ -3,7 +3,14 @@
 namespace RonasIT\Larabuilder\Visitors;
 
 use PhpParser\Node;
+use PhpParser\Node\Arg;
+use PhpParser\Node\ArrayItem;
 use PhpParser\Node\Expr\Array_;
+use PhpParser\Node\Expr\ConstFetch;
+use PhpParser\Node\Name;
+use PhpParser\Node\Scalar\Float_;
+use PhpParser\Node\Scalar\Int_;
+use PhpParser\Node\Scalar\String_;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\ClassConst;
 use PhpParser\Node\Stmt\ClassMethod;
@@ -78,5 +85,49 @@ abstract class BaseNodeVisitorAbstract extends NodeVisitorAbstract
                 }
             }
         }
+    }
+
+    protected function makeArg(mixed $value): Arg
+    {
+        list($value) = $this->getPropertyValue($value);
+
+        return new Arg($value);
+    }
+
+    protected function getPropertyValue(mixed $value): array
+    {
+        $type = get_debug_type($value);
+
+        $value = match ($type) {
+            'int' => new Int_($value),
+            'array' => $this->makeArrayValue($value),
+            'string' => new String_($value),
+            'float' => new Float_($value),
+            'bool' => $this->makeBoolValue($value),
+            'null' => new ConstFetch(new Name('null')),
+        };
+
+        return [$value, $type];
+    }
+
+    protected function makeBoolValue(bool $value): ConstFetch
+    {
+        $name = new Name(($value) ? 'true' : 'false');
+
+        return new ConstFetch($name);
+    }
+
+    protected function makeArrayValue(array $values): Array_
+    {
+        $items = [];
+
+        foreach ($values as $key => $val) {
+            list($val) = $this->getPropertyValue($val);
+            list($key) = $this->getPropertyValue($key);
+
+            $items[] = new ArrayItem($val, $key);
+        }
+
+        return new Array_($items);
     }
 }
