+++ tests/AppBootstrapBuilderTest.php
@@ -131,7 +131,7 @@ class AppBootstrapBuilderTest extends TestCase
         );
 
         new AppBootstrapBuilder()
-            ->addScheduleCommand(
+            ->addSchedule(
                 'telescope:prune --set-hours=resolved_exception:1,completed_job:0.1 --hours=336',
                 'production',
                 new ScheduleFrequencyOptionsDTO(
@@ -142,7 +142,7 @@ class AppBootstrapBuilderTest extends TestCase
                     attributes: ['America/New_York'],
                 ),
             )
-            ->addScheduleCommand(
+            ->addSchedule(
                 command: 'telescope:prune --set-hours=resolved_exception:12222',
             )
             ->save();
@@ -157,11 +157,11 @@ class AppBootstrapBuilderTest extends TestCase
         );
 
         new AppBootstrapBuilder()
-            ->addScheduleCommand(
+            ->addSchedule(
                 command: 'telescope:prune --set-hours=resolved_exception:1,completed_job:0.1 --hours=336',
                 environment: 'production',
             )
-            ->addScheduleCommand(
+            ->addSchedule(
                 command: 'telescope:prune --set-hours=resolved_exception:12222',
             )
             ->save();
@@ -176,7 +176,7 @@ class AppBootstrapBuilderTest extends TestCase
         );
 
         new AppBootstrapBuilder()
-            ->addScheduleCommand(
+            ->addSchedule(
                 command: 'telescope:prune --set-hours=resolved_exception:1,completed_job:0.1 --hours=336',
                 environment: 'production',
             )
@@ -184,14 +184,14 @@ class AppBootstrapBuilderTest extends TestCase
                 exceptionClass: HttpException::class,
                 renderBody: 'return;',
             )
-            ->addScheduleCommand(
+            ->addSchedule(
                 command: 'telescope:prune --set-hours=resolved_exception_2',
             )
             ->addExceptionsRender(
                 exceptionClass: HttpException::class,
                 renderBody: 'return;',
             )
-            ->addScheduleCommand(
+            ->addSchedule(
                 command: 'telescope:prune --set-hours=resolved_exception_3',
             )
             ->save();
